Add files via upload
This implementation includes basic mehods as insert, delete, search and print, as well as one additional function that creates a dictionary of words as the Key value and amount of occurances of its Key as an Info value from a .txt file. And also 2 operators(= and [ ]). This header might be used as a help for creating custom hashtable to reach logN complexity of search function.